Put SolvatedPDBComponent into separate file
