+++ gufe/components/__init__.py
@@ -1,6 +1,7 @@
 """The building blocks for defining systems"""
 
 from .component import Component
-from .proteincomponent import ProteinComponent, ProteinMembraneComponent, SolvatedPDBComponent
+from .proteincomponent import ProteinComponent
 from .smallmoleculecomponent import SmallMoleculeComponent
 from .solventcomponent import BaseSolventComponent, SolventComponent
+from .solvatedpdbcomponent import SolvatedPDBComponent, ProteinMembraneComponent
